+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changes
 
 ## February 17, 2025
+- Added loading indicator when expanding directories in the file browser
 - Added WAV file plugin with audio playback and waveform visualization
 - Updated URL query parameter mapping between v1 to v2
 - Adjusted home page layout